backend/requirements.txt
# Core application dependecies
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
requests

# Database
sqlalchemy
psycopg2-binary

# Security & Authentication
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
